refactor axecore check into helper function
